Implement several helper functions
Add playerSelection, computerSelection and playRound function so a full
game of rock paper scissors can be played.

Add playerSelection to get a valid input for rock, paper or scissors.
Add computerSelection to get a valid input for rock, paper or scissors
from a random number (1 to 3).
Add playRound to play a full round of rock paper scissors with two
arguements, player- and computerchoice and return the output as string.